backend/requirements.txt
markdown
django==4.2.11
gunicorn==22.0.0
uvicorn[standard]==0.29.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
django-filter==24.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
pillow==10.3.0
django-db-connection-pool[postgresql]==1.2.5
ulid-py==1.1.0
drf-spectacular==0.27.2
django-redis==5.4.0
celery==5.4.0
